requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.6
openai>=1.40.0
python-dotenv>=1.0.1

aiohttp
psycopg2-binary
SQLAlchemy>=2.0
pgvector>=0.2.5
sentence-transformers
pypdf
python-docx
python-multipart

structlog
colorama